Internacionalización

+++ src/app/pages/movies/show/show-movie.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { TranslateService } from '@ngx-translate/core';
 import { Movie } from 'src/app/@core/model/Movie';
 import { ActorsService } from 'src/app/@core/services/actors.service';
+import { AlertMessage } from 'src/app/@core/services/AlertMessage';
 import { MoviesService } from 'src/app/@core/services/movies.service';
 import { ToastMessage } from 'src/app/@core/services/ToastMessage';
 import { UtilService } from 'src/app/@core/services/util.service';
@@ -14,7 +16,7 @@ import { UtilService } from 'src/app/@core/services/util.service';
 export class ShowMovieComponent implements OnInit {
 
   public id: string;
-  public movie: Movie = new Movie();
+  public movie: Movie;
   public loading: boolean;
   public actors: any[] = [];
 
@@ -22,7 +24,8 @@ export class ShowMovieComponent implements OnInit {
     private moviesService: MoviesService,
     private utilService: UtilService,
     private actorsService: ActorsService,
-    private router: Router) { }
+    private router: Router,
+    public translate: TranslateService) { }
 
   ngOnInit() {
     this.loading = true;
@@ -31,28 +34,47 @@ export class ShowMovieComponent implements OnInit {
   }
 
   async getMovieByID() {
-    let res = await this.moviesService.getMovieByID(this.id);
-    if (res) {
-      this.movie = res.body;
-      let actors: [] = (await this.actorsService.getActors()).body;
-      this.movie.actors.forEach(a => this.actors.push(actors.find((act: any) => act.id == a)));
-    }
-    else
-      this.utilService.showToast(ToastMessage.ErrorGeneric);
-    this.loading = false;
+    await this.moviesService.getMovieByID(this.id)
+      .then(async (res) => {
+        this.movie = res.body;
+        this.getActorsByMovie();
+      })
+      .catch((error) => {
+        this.loading = false
+        this.utilService.showToast(ToastMessage.ShowError, this.translate.instant('edit-movie.show-error'));
+      });
+  }
+
+  async getActorsByMovie() {
+    await this.actorsService.getActors()
+      .finally(() => this.loading = false)
+      .then(async (res) => {
+        let actors = res.body;
+        this.movie.actors.forEach(a => this.actors.push(actors.find((act: any) => act.id == a)));
+      })
+      .catch((error) => {
+        this.utilService.showToast(ToastMessage.ShowError, this.translate.instant('edit-movie.show-error'));
+      });
   }
 
   async deleteMovie() {
-    this.loading = true;
-    await this.moviesService.deleteMovie(this.id)
-      .finally(() => this.loading = false)
-      .then(async res => {
-        this.router.navigate(['/pages/movies/list']);
-        this.utilService.showToast(ToastMessage.DeleteOK, 'Película');
+    await this.utilService.message(AlertMessage.GenericQuestion, this.translate.instant('sweet-alert.delete-movie'))
+      .then(async (value) => {
+        if (value.isConfirmed) {
+          this.loading = true;
+          await this.moviesService.deleteMovie(this.id)
+            .finally(() => this.loading = false)
+            .then(async res => {
+              this.router.navigate(['/pages/movies/list']);
+              this.utilService.showToast(ToastMessage.DeleteOK, this.translate.instant('toast.movie'));
+            })
+            .catch((error) => {
+              this.utilService.showToast(ToastMessage.ShowError, this.translate.instant('edit-movie.show-error'));
+            });
+        }
       })
       .catch((error) => {
-        this.utilService.showToast(ToastMessage.ErrorGeneric);
-        console.log(error);
+        this.utilService.showToast(ToastMessage.ShowError, this.translate.instant('edit-movie.show-error'));
       });
   }
 }
